feat: splits the emailSettingsModel file into two seperate files
